[OTServ Trunk] * Added conditions code from avesta * onStepIn and onStepOut events are now executed on the correct order * As of Tibia 7.9, spawns aren't affected by players on lower or higher floors of them
[OTServ Avesta]
* Updated avesta to the most recent OTServ trunk revision(including this rev)
* Remove compiler warnings

[Data Trunk]
* Swimming and Diving systems - ATTENTION! Before trying these codes, read the INSTRUCTIONS file on the documents folder!
* Deleted TODO - you can view it on http://otserv.org/index.php?title=Scripts_Roadmap

[All three]
* Fixes and sync items.xml

+++ otserv/branches/avesta/md5.h
@@ -52,9 +52,11 @@
 #ifndef ___MD5_H___
 #define ___MD5_H___
 
+#include "definitions.h"
+
 /* Typedef a 32 bit type */
 #ifndef UINT4
-typedef unsigned long int UINT4;
+typedef uint32_t UINT4;
 #endif
 
 /* Data structure for MD5 (Message Digest) computation */
